Remove warnings on core and tests

+++ core/index3d.c
@@ -46,15 +46,15 @@ void *index3d_get(const Index3D *index, const int32_t x, const int32_t y, const 
     static uint32_t uy;
     static uint32_t uz;
 
-    ux = x;
-    uy = y;
-    uz = z;
+    ux = (uint32_t)x;
+    uy = (uint32_t)y;
+    uz = (uint32_t)z;
 
     //    static void** currentNode;
     void **currentNode = index->topLevelNode;
 
-    static int32_t modulo;
-    static int32_t quotient;
+    static uint32_t modulo;
+    static uint32_t quotient;
     modulo = ux & INDEX_NODE_BITWISE_MODULO;
     quotient = ux >> INDEX_DIVIDE_BYTES;
 
@@ -160,14 +160,14 @@ void *index3d_remove(Index3D *index,
                      Index3DIterator *it) {
 
     // only use unsigned integers for bitwise operations:
-    uint32_t ux = x;
-    uint32_t uy = y;
-    uint32_t uz = z;
+    uint32_t ux = (uint32_t)x;
+    uint32_t uy = (uint32_t)y;
+    uint32_t uz = (uint32_t)z;
 
     void **currentNode = index->topLevelNode;
 
-    int32_t modulo = ux & INDEX_NODE_BITWISE_MODULO;
-    int32_t quotient = ux >> INDEX_DIVIDE_BYTES;
+    uint32_t modulo = ux & INDEX_NODE_BITWISE_MODULO;
+    uint32_t quotient = ux >> INDEX_DIVIDE_BYTES;
 
     // look for x
 
@@ -283,14 +283,14 @@ void index3d_insert(Index3D *index,
                     Index3DIterator *it) {
 
     // only use unsigned integers for bitwise operations:
-    uint32_t ux = x;
-    uint32_t uy = y;
-    uint32_t uz = z;
+    uint32_t ux = (uint32_t)x;
+    uint32_t uy = (uint32_t)y;
+    uint32_t uz = (uint32_t)z;
 
     void **currentNode = index->topLevelNode;
 
-    int32_t modulo = ux & INDEX_NODE_BITWISE_MODULO;
-    int32_t quotient = ux >> INDEX_DIVIDE_BYTES;
+    uint32_t modulo = ux & INDEX_NODE_BITWISE_MODULO;
+    uint32_t quotient = ux >> INDEX_DIVIDE_BYTES;
 
     // set x
